requirements.txt
Django==4.1.7
dj-database-url==1.2.0
whitenoise==6.4.0
gunicorn==20.1.0
psycopg2==2.9.5
Pillow==9.4.0
django-allauth==0.52.0
django-social-share==2.3.0
stripe==5.4.0

# Google
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.58.0
google-api-core==2.11.0
google-api-python-client==2.80.0
oauthlib==3.2.2

# Config
python-dotenv==1.0.0
